crypto: msm: Make changes to the reaper timer
Change the periodical reaper timer to on-demand basis. The reaper timer
is only set to disable clock after the clock is enabled, and it is
removed when there is no crypto operations. So that we can prevent the
timers from unnecessarily waking up the CPU from power collapse. The
timer interval is changed to 1 second.

Signed-off-by: Zhen Kong <zkong@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>